+++ routes/site/site.js
@@ -13,6 +13,7 @@ const siteMw            = require('../../middleware/site');
 const voteMw            = require('../../middleware/vote');
 const userClientMw      = require('../../middleware/userClient');
 const roleClientMw      = require('../../middleware/role');
+const requestSiteValidation = require('../../middleware/request/validation/site');
 
 //services
 const userClientApi     = require('../../services/userClientApi');
@@ -58,7 +59,7 @@ const formatDomainFromBody = (req, res, next) => {
     domain = req.body.domain
   }
 
-  const protocol = req.body.protocol ? req.body.protocol : 'https://';
+  const protocol = req.body.protocol || ( process.env.FORCE_HTTP ? 'http://' : 'https://' );
   domain = protocol + cleanUrl(domain);
   req.formattedDomain = domain;
 
@@ -208,12 +209,16 @@ module.exports = function(app){
           }
 
           req.flash('success', { msg: 'De site is succesvol aangemaakt'});
-        res.redirect('/admin/site/' + site.id)
+          req.session.save( () => {
+              res.redirect('/admin/site/' + site.id)
+          });
 
       } catch (error) {
         console.error(error);
         req.flash('error', { msg: 'Het is helaas niet gelukt om de site aan te maken.'});
-        res.redirect('back');
+        req.session.save( () => {
+          res.redirect('back');
+        });
       }
     }
   );
@@ -264,11 +269,15 @@ module.exports = function(app){
           }
 
         req.flash('success', { msg: 'De site is succesvol aangemaakt'});
-        res.redirect('/admin/site/' + site.id);
+        req.session.save( () => {
+          res.redirect('/admin/site/' + site.id);
+        });
       } catch(error) {
         console.error(error);
         req.flash('error', { msg: 'Het is helaas niet gelukt om de site aan te maken'});
-        res.redirect('back');
+        req.session.save( () => {
+          res.redirect('back');
+        });
       }
     },
   );
@@ -324,7 +333,9 @@ module.exports = function(app){
         console.log('SITE EXPORT FAILED');
         console.log(err);
         req.flash('error', { msg: 'Het is helaas niet gelukt om de site te exporteren'});
-        res.redirect('back');
+        req.session.save( () => {
+          res.redirect('back');
+        });
       }
 
     }
@@ -336,6 +347,7 @@ module.exports = function(app){
   app.post('/admin/site/:siteId(\\d+)',
     siteMw.withOne,
     siteMw.withAll,
+    requestSiteValidation,
     (req, res, next) => {
       delete req.body.url;
       const siteConfigFields = Object.keys(siteConfigSchema);
@@ -401,7 +413,9 @@ module.exports = function(app){
 
           await k8Ingress.ensureIngressForAllDomains(req.sites)
 
-          res.redirect(req.header('Referer')  || appUrl);
+          req.session.save( () => {
+            res.redirect(req.header('Referer')  || appUrl);
+          });
         })
         .catch((err) => { next(err) });
     }
@@ -472,12 +486,16 @@ module.exports = function(app){
 
 
           req.flash('success', { msg: 'Url aangepast!'});
-          res.redirect(req.header('Referer')  || appUrl);
+          req.session.save( () => {
+            res.redirect(req.header('Referer')  || appUrl);
+          });
         })
         .catch(function (err) {
           console.error(err);
           req.flash('error', { msg: 'Er gaat iets mis!'});
-          res.redirect(req.header('Referer')  || appUrl);
+          req.session.save( () => {
+            res.redirect(req.header('Referer')  || appUrl);
+          });
         });
 
     }
@@ -513,12 +531,16 @@ module.exports = function(app){
       Promise.all(promises)
         .then(function (response) {
           req.flash('success', { msg: 'Aangepast!'});
-          res.redirect(req.header('Referer')  || appUrl);
+          req.session.save( () => {
+            res.redirect(req.header('Referer')  || appUrl);
+          });
         })
         .catch(function (err) {
           console.error(err);
           req.flash('error', { msg: 'Er gaat iets mis!'});
-          res.redirect(req.header('Referer')  || appUrl);
+          req.session.save( () => {
+            res.redirect(req.header('Referer')  || appUrl);
+          });
         });
 
     }
@@ -540,14 +562,18 @@ module.exports = function(app){
       Promise.all(anonymizeActions)
         .then((response) => {
           req.flash('success', { msg: 'Geannonimiseerd!'});
-          res.redirect('/admin');
+          req.session.save( () => {
+            res.redirect('/admin');
+          });
         })
         .catch((err) => {
           console.log('ANONYMIZE WEBSITE FAILED');
           console.log(err);
           let message = err && err.error && err.error.message ?  'Er gaat iets mis: '+ err.error.message : 'Er gaat iets mis!';
           req.flash('error', { msg: message});
-          res.redirect('/admin');
+          req.session.save( () => {
+            res.redirect('/admin');
+          });
         });
     }
   );
@@ -582,14 +608,18 @@ module.exports = function(app){
         })
         .then( result => {
           req.flash('success', { msg: 'Verwijderd!'});
-          res.redirect('/admin');
+          req.session.save( () => {
+            res.redirect('/admin');
+          });
         })
         .catch( err => {
           console.log('DELETE WEBSITE FAILED');
           console.log(err);
           let message = err && err.error && err.error.message ?  'Er gaat iets mis: '+ err.error.message : 'Er gaat iets mis!';
           req.flash('error', { msg: message});
-          res.redirect('/admin');
+          req.session.save( () => {
+            res.redirect('/admin');
+          });
         });
     }
   );