+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## UNRELEASED
+## v0.41.0
 * Use networking.k8s.io/v1 version for ingress
   * IMPORTANT: This requires Kubernetes version v1.19 due to switching from networking.k8s.io/v1beta1 to networking.k8s.io/v1. 
   * See https://kubernetes.io/docs/reference/using-api/deprecation-guide/#ingress-v122 for more details.